Start with openafs 1.8.pre1 as basis
This is just the original openafs 1.8.pre1 without any OSD extensions.
They will follow in other commits.

+++ src/WINNT/afsd/cm_access.c
@@ -7,7 +7,10 @@
  * directory or online at http://www.openafs.org/dl/license10.html
  */
 
+#include <afsconfig.h>
 #include <afs/param.h>
+#include <roken.h>
+
 #include <afs/stds.h>
 
 #include <windows.h>
@@ -30,44 +33,49 @@ int cm_accessPerFileCheck = 0;
  * can't be locked.  Thus, this must always be called in a while loop to stabilize
  * things, since we can always lose the race condition getting to the parent vnode.
  */
-int cm_HaveAccessRights(struct cm_scache *scp, struct cm_user *userp, afs_uint32 rights,
+int cm_HaveAccessRights(struct cm_scache *scp, struct cm_user *userp, cm_req_t *reqp,
+                        afs_uint32 rights,
                         afs_uint32 *outRightsp)
 {
     cm_scache_t *aclScp;
-    long code;
+    long code = 0;
     cm_fid_t tfid;
     int didLock;
     long trights;
     int release = 0;    /* Used to avoid a call to cm_HoldSCache in the directory case */
+    cm_volume_t *volp = cm_FindVolumeByFID(&scp->fid, userp, reqp);
 
     didLock = 0;
-    if (scp->fileType == CM_SCACHETYPE_DIRECTORY || cm_accessPerFileCheck) {
+    if (scp->fileType == CM_SCACHETYPE_DIRECTORY ||
+        cm_accessPerFileCheck ||
+        !volp || (volp->flags & CM_VOLUMEFLAG_DFS_VOLUME)) {
         aclScp = scp;   /* not held, not released */
     } else {
         cm_SetFid(&tfid, scp->fid.cell, scp->fid.volume, scp->parentVnode, scp->parentUnique);
         aclScp = cm_FindSCache(&tfid);
-        if (!aclScp)
-            return 0;
+        if (!aclScp) {
+            code = 0;
+            goto done;
+        }
+        release = 1;
         if (aclScp != scp) {
             if (aclScp->fid.vnode < scp->fid.vnode)
                 lock_ReleaseWrite(&scp->rw);
-            lock_ObtainRead(&aclScp->rw);
+            lock_ObtainWrite(&aclScp->rw);
+	    didLock = 1;
             if (aclScp->fid.vnode < scp->fid.vnode)
                 lock_ObtainWrite(&scp->rw);
 
-	    /* check that we have a callback, too */
+            /* check that we have a callback, too */
             if (!cm_HaveCallback(aclScp)) {
                 /* can't use it */
-                lock_ReleaseRead(&aclScp->rw);
-                cm_ReleaseSCache(aclScp);
-                return 0;
+                code = 0;
+                goto done;
             }
-            didLock = 1;
         }
-        release = 1;
     }
 
-    lock_AssertAny(&aclScp->rw);
+    lock_AssertWrite(&aclScp->rw);
 
     /* now if rights is a subset of the public rights, we're done.
      * Otherwise, if we an explicit acl entry, we're also in good shape,
@@ -105,7 +113,7 @@ int cm_HaveAccessRights(struct cm_scache *scp, struct cm_user *userp, afs_uint32
 		      scp, scp->unixModeBits);
 	    *outRightsp &= ~PRSFS_WRITE;
 	}
-	if ((scp->unixModeBits & 0200) == 0 && !cm_deleteReadOnly) {
+	if ((scp->unixModeBits & 0200) == 0 && !cm_deleteReadOnly && (reqp->flags & CM_REQ_SOURCE_SMB)) {
 	    osi_Log2(afsd_logp,"cm_HaveAccessRights UnixMode removing DELETE scp 0x%p unix 0%o",
 		      scp, scp->unixModeBits);
 	    *outRightsp &= ~PRSFS_DELETE;
@@ -129,8 +137,10 @@ int cm_HaveAccessRights(struct cm_scache *scp, struct cm_user *userp, afs_uint32
     /* fall through */
 
   done:
+    if (volp)
+        cm_PutVolume(volp);
     if (didLock)
-        lock_ReleaseRead(&aclScp->rw);
+        lock_ReleaseWrite(&aclScp->rw);
     if (release)
         cm_ReleaseSCache(aclScp);
     return code;
@@ -150,6 +160,7 @@ long cm_GetAccessRights(struct cm_scache *scp, struct cm_user *userp,
     cm_fid_t tfid;
     cm_scache_t *aclScp = NULL;
     int got_cb = 0;
+    cm_volume_t * volp = cm_FindVolumeByFID(&scp->fid, userp, reqp);
 
     /* pretty easy: just force a pass through the fetch status code */
 
@@ -158,18 +169,21 @@ long cm_GetAccessRights(struct cm_scache *scp, struct cm_user *userp,
     /* first, start by finding out whether we have a directory or something
      * else, so we can find what object's ACL we need.
      */
-    if (scp->fileType == CM_SCACHETYPE_DIRECTORY || cm_accessPerFileCheck) {
+    if (scp->fileType == CM_SCACHETYPE_DIRECTORY ||
+        cm_accessPerFileCheck ||
+        !volp || (volp->flags & CM_VOLUMEFLAG_DFS_VOLUME))
+    {
 	code = cm_SyncOp(scp, NULL, userp, reqp, 0,
 			 CM_SCACHESYNC_NEEDCALLBACK | CM_SCACHESYNC_GETSTATUS | CM_SCACHESYNC_FORCECB);
 	if (!code)
 	    cm_SyncOpDone(scp, NULL, CM_SCACHESYNC_NEEDCALLBACK | CM_SCACHESYNC_GETSTATUS);
-    else
-        osi_Log3(afsd_logp, "GetAccessRights syncop failure scp %x user %x code %x", scp, userp, code);
+        else
+            osi_Log3(afsd_logp, "GetAccessRights syncop failure scp %x user %x code %x", scp, userp, code);
     } else {
         /* not a dir, use parent dir's acl */
         cm_SetFid(&tfid, scp->fid.cell, scp->fid.volume, scp->parentVnode, scp->parentUnique);
         lock_ReleaseWrite(&scp->rw);
-        code = cm_GetSCache(&tfid, &aclScp, userp, reqp);
+        code = cm_GetSCache(&tfid, NULL, &aclScp, userp, reqp);
         if (code) {
             lock_ObtainWrite(&scp->rw);
 	    goto _done;
@@ -190,5 +204,7 @@ long cm_GetAccessRights(struct cm_scache *scp, struct cm_user *userp,
     }
 
   _done:
+    if (volp)
+        cm_PutVolume(volp);
     return code;
 }
